[IMP] website, *: make 403 error page editable via website editor
*= http_routing
This commit allows users to customize the 403 (Forbidden) error page
directly from the website builder by introducing a dedicated route:
`/website/http_error/403`. This enables full editing of the page using
snippets and inner blocks, and provides the ability to toggle the
visibility of the default error message, which is shown when a 403
error is encountered.
A generic controller is also added to provide access to all available
error pages.

task-4714826