[IMP] html_builder, website: use arrow keys to navigate in tab panes
The tab panes of the AddSnippetDialog and the AddPageDialog, as well as
the snippets blocks in the sidebar, would benefit from a proper keyboard
navigation. Since [9ae02d8], the snippets and page layouts are tabable.
But when there are many snippets or layouts, an approach using the
arrows would be more practical and intuitive, as the dialogs and the
snippets categories all use a masonry layout.

The logic makes sure to handle both LTR and RTL languages.

[9ae02d8]: https://github.com/odoo/odoo/commit/9ae02d80894d4043e49d8e2cad068f8018e6f113

task-5384672