[FIX] hr{,_presence}: employee onboarding helper
Issue:
------

The "employee no content onboarding helper" (screen with `Load sample data` button) shouldn't be shown on other companies `Employee` app than the `base.main_company`. Indeed, this button only loads the sample data for `base.main_company`.

Reproducing steps:
------------------

- Create a DB with `hr` module and without demo data
- Create a second company
- Go in the Employees app, the onboarding helper should be displayed
- Click on the "Load sample data" button, the page is reloaded, but nothing happens

Purpose:
--------

Improving the kanban no content employee helper so that :
A) It displays "Load sample data" and "Create an employee" buttons with some onboarding design things when the user has the rights to create employees, one of the displayed company is `base.main_company` and the displayed companies have no employees.
B) If all the condition of A are validated, except the user hasn't the employee rights, then display the onboarding things without the 2 buttons.
C) Otherwise same behavior as now (display default helper or custom helper defined from the view)

Same for mobile, but the (A) view comes with a simpler design if there is not enough space.

community-229238
task-4879557