[IMP] mail: introduce call artifacts, call_debrief player widget
This commit introduces a unified way to store and visualize post-call
artifacts (call recordings).
As well as it lays groundwork for introducing transcripts. Primarily for Discuss and VoIP.

CALL ARTIFACT - what is that?
---
A call.artifact represents a discrete "product" of a call. For exmple it can be a call's recording or call transcription.
Each artifact acts as a thin metadata layer (`start_ms`, `end_ms`, `source`) for one `ir.attachment` at maximum.
This keeps the generic `ir.attachment` table lean while allowing us to store complex, multi-segment timing metadata,
required for chunked playback and AI transcription pipelines.

RECORDING STORAGE
---
Odoo's infrastructure is not designed for mass storage of media (ref.1).
Therefore, the persistent call recording feature will ONLY be available when the
cloud storage module is installed and configured allowing media to be offloaded to providers like S3 or GCS.

Note: Transient recordings used for transcriptions are not bounded by this requirement.
(Check out the enterprise commit for more details.)

Call Debrief - NEW UI MEDIA PLAYER
---
Presentation of the artifacts is realized with Call Debrief Widget - a new
field widget. It provides a modern synchronized player supporting audio/video
recordings (with support for synchronized transcripts), drag-to-seek, keyboard
shortcuts with live feedback and variable playback rates.

References
---
ref.1 Even tho a single recording size is negligible, we have to keep in
mind that the voip systems are expected to generate massive numbers of
them.

task-5153790