[FIX] purchase_stock: show supply warning correctly in multi-company
Issue:
------------------------------------------
In a `multi-company` environment, if supplier `price rules` exist in one company
but not in another, creating a `reordering rule` in the company without any
price rules still shows the `info icon` instead of the `warning icon`.

How to reproduce:
------------------------------------------
1. Install Stock and Purchase.
2. Create a product P1.
3. Add a supplier under Company A.
4. Switch to Company B.(in multi-company env)
5. Create a reordering rule for product P1.
The reordering rule incorrectly shows the `info icon` even though `no supplier`
exists in `Company B`.

Cause of the issue:
------------------------------------------
The computation of `show_supply_warning` only checks whether `suppliers` exist
for the `product`, without filtering them by company of orderpoint.

Solution:
------------------------------------------
Used `allowed_partner_ids` to filter `suppliers` by the `company` of the
`reordering rule` and check whether any supplier exists or not.

This helps users to correctly see a warning icon in rr to easily find
the rr with missing company-specific vendor setups in it.

TaskId: 5047947