[IMP] mail,*: preserve the message formatting in various message preview
*=html_editor

Before this commit,
The message preview that was shown in various parts of the applications shows
only the inline version of the message with formatting being stripped out.
This gave false impression about how the original message actually is...

This commit, retains the message's original formatting and making it inline
by adding appropriate spaces in place of <br> and in between two block level
elements.

Task-5263284