[IMP] html_editor: split HistoryPlugin DomMutationPlugin + HistoryPlugin
[FIX] html_editor: adapt batched steps mechanism to new split

[FIX] html_editor: do not revert the initial step

[FIX] html_editor: fix order of operations on commit and start editor

[FIX] html_editor: restore return guard in restoreToStep

[FIX] html_editor: restore history reset order

[FIX] html_editor: no reset timestamp on add external step

[FIX] html_editor: do not pass commit root ID as it might not exist

If a node is not observed it won't have an ID. Then passing the root's
ID like that will fail. This was breaking all AI
voice_transcription_plugin tests.

[FIX] html_editor: adjust on_editor_started_handlers sequence

`on_editor_started_handlers` has to trigger in `translation`` first,
then in `domMutation`, then in `history`. Now I had a `withSequence` of
0 on `domMutation` to make sure it triggers before `history`, but
`translation` having a sequence of 5 triggers after `domMutation`. I'll
change the sequence to 9 so it's after `translation` and since they're
the only two with a sequence they'll be the first two (the default
sequence is 10). `translate attribute history` passes now and the rest
should remain good.

[FIX] html_editor: reset `domMutation.currentChanges` after each commit

This fixes the tours:
- `website_update_column_count`
- `edit_megamenu_visibility`
- `website_form_editor_tour`

which failed because of the adaptation of `extraStepInfos`. Master
resets `currentStep` after adding a step, but in this branch,
`domMutation` wasn't resetting its equivalent (`currentChanges`) fully,
so the extra step infos remained and it was stale. As a result,
`builder_options_plugin` had the wrong `currentTarget` and/or
`nextTarget` values.

[REF] mass_mailing: remove unnecessary historyPlugin fetch

Was added by commit [1]. Should have probably been removed by commit
[2].

[1]:
    https://github.com/odoo/odoo/commit/438ec1aafd3120483a3166084e71a8069f65b497
[2]:
    https://github.com/odoo/odoo/commit/939088eb9228126cca818f2b29ae354e8c4bc0fb

[REF] html_editor, *: merge steps with commits

This merges the concepts of step and commit together.

[REF] html_editor: move batchable and commitTimestamp to metadata

[REF] html_editor: extract currentChanges phase 1

[REF] html_editor: proposal for private current changes

[REF] html_editor: remove unnecessary condition

[REF] html_editor: move processNewRecords call inside condition

[REF] html_editor: move comments around

[REF] html_editor: review stage/handle flow

[REF] html_editor: review stage/handle flow PHASE 2

[REF] html_editor: merge stage functions PHASE 1

[REF] html_editor: merge stage functions PHASE 2: merge stage and flush