[FIX] sale{,_purchase}: correct SOL description display when adding a product
Steps to reproduce:
---------------------
1. Install the Sales and Purchase modules.
2. Create a product, then go to the Purchase tab and add/create a vendor with
a vendor product name and code.
3. Create a Sale Order using this vendor as the customer, and add
the product to the sale order line.
4. Observe the behavior of the line before and after clicking outside the field.

Observation and issue:
------------------------
1. Before clicking outside the SOL, the description remains
unfilled despite the availability of a vendor product name and code.

2. Upon clicking outside the SOL, the description displays
only the vendor product name and code.

Cause:
---------
- The `partner_id` in the context (from `get_single_product_variant`
  and `_compute_translated_product_name`) causes the product
  `display_name` to be computed using vendor information.
https://github.com/odoo/odoo/blob/c3aa5a9c3031823baf98a395a3383ced5e07e9e7/addons/product/models/product_product.py#L538-L548

- The SOL `_compute_name` relies on the product `display_name`,
but missing dependency values in the context prevent `_compute_display_name`
from being recomputed

Solution:
------------
Ensure the SOL description  correctly incorporates the product description
when vendor product_name and product_code are present. Add the missing
dependency values in the context so that the SOL description is recomputed.

opw-5262670