[IMP] mrp,*: move compute price from BoM action to BoM list view
*: mrp_{account, subcontracting, subcontracting_account}, sale_mrp_{margin},
purchase_mrp

Currently, the `Compute price from BoM` action is available directly on the
product form. Because it is highly visible and easily accessible, users tend to
trigger it frequently, even when there is no need to recompute the cost.

Moreover, in `AVCO and FIFO` cost configurations, manually updating the product
cost based on the BoM is conceptually incorrect. Triggering this action can lead
to misleading cost values and may cause inventory valuation or accounting
inconsistencies.

Additionally, the BoM cost computation action was defined in the `mrp_account`
module even though it only computes production cost based on components and
operations defined in the BoM. This creates an unnecessary dependency on
`mrp_account` module and prevents `mrp` only users from accessing the feature.

This PR moves cost computation actions away from the product form and places
them at the BoM level, which is the actual source of truth for this information.

- The `Compute price from BoM` action has been removed from the product form
and list view to prevent accidental or repetitive usage. A new `Update Product
Cost from BoM` action is made available directly on the BoM list view and the
related computation logic has been moved from `mrp_account` to `mrp` to module
to remove the unnecessary dependency.
- The `Unit Cost` is now displayed by default in the BoM list view, providing
better visibility into BoM-driven costs.

Overall, this improvement minimizes unintended cost updates, and ensures that
BoM-based cost computation is used only when it is functionally relevant,
resulting in safer and more predictable accounting behavior.

TaskID-5180166